Refactor React to have multiple pages
