+++ src/App.css
@@ -37,7 +37,7 @@
   display: none;
   list-style: none;
   padding: 10px;
-  margin: 5px 0px;
+  margin: 6px 0px;
   height: 0px;
   right: 0;
   overflow: hidden;
